+++ android/src/main/java/com/getcapacitor/community/database/sqlite/SQLite/UtilsSecret.java
@@ -167,4 +167,11 @@ public class UtilsSecret {
     public void clearPassphrase() {
         sharedPreferences.edit().remove("secret").commit();
     }
+
+    public Boolean isPassphrase() {
+        if (!getPassphrase().isEmpty()) {
+            return true;
+        }
+        return false;
+    }
 }